src-query/consumer/requirements.txt
kafka-python==2.3.0
redis==7.2.0
pydantic==2.12.5
python-dotenv==1.2.1